[HEAP-8587] Fix sessionization for the `TestDriver` android app.
It turns out that `Heap.init` is called (from Javascript) too late,
after `MainActivity` is created. This leads our library to assume that
`MainActivity` is backgrounded, and so all events are treated as custom
events, with each receiving a different session ID.

The fix is simple; call `Heap.init` from `MainActivity` directly (in
Java), before the Javascript is loaded.